Fix an important bug on how we create the states.
There was a corner case when if the time between the first flow and the
second flow was more than the timeout threshold, then we were not adding
the corresponding 0s. Now we are.

I'm not sure if we should regerenate much stuff, because the vast
mayority of the models (99.9%?) are not affected.